app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythondi==1.2.4
regex==2024.11.6
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
watchfiles==1.0.3
